Changed app name and packages Modified colors Added toolbar with back button
