--- app/src/androidTest/java/yt/javi/nfltweets/ExampleInstrumentedTest.kt
@@ -1,24 +1,0 @@
-package yt.javi.nfltweets
-
-import android.support.test.InstrumentationRegistry
-import android.support.test.runner.AndroidJUnit4
-import org.junit.Assert.assertEquals
-import org.junit.Test
-import org.junit.runner.RunWith
-
-/**
- * Instrumentation test, which will execute on an Android device.
- *
- * @see [Testing documentation](http://d.android.com/tools/testing)
- */
-@RunWith(AndroidJUnit4::class)
-class ExampleInstrumentedTest {
-    @Test
-    @Throws(Exception::class)
-    fun useAppContext() {
-        // Context of the app under test.
-        val appContext = InstrumentationRegistry.getTargetContext()
-
-        assertEquals("yt.javi.nfltweets", appContext.packageName)
-    }
-}
