+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="yt.javi.nfltweets">
+    package="yt.javi.nftweets">
     <uses-permission android:name="android.permission.INTERNET" />
 
     <application
@@ -10,14 +10,14 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".MainActivity">
+        <activity android:name="yt.javi.nftweets.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".TwitterTimeLineActivity"></activity>
+        <activity android:name="yt.javi.nftweets.TwitterTimeLineActivity"></activity>
     </application>
 
 </manifest>
