--- app/src/main/java/yt/javi/nfltweets/domain/infrastructure/repositories/inmemory/TeamInMemoryRepository.kt
@@ -1,15 +1,0 @@
-package yt.javi.nfltweets.domain.infrastructure.repositories.inmemory
-
-import yt.javi.nfltweets.domain.model.Conference
-import yt.javi.nfltweets.domain.model.Division
-import yt.javi.nfltweets.domain.model.team.Team
-import yt.javi.nfltweets.domain.model.team.TeamRepository
-
-
-class TeamInMemoryRepository(private val teams: List<Team>): TeamRepository {
-    override fun findTeams(conference: Conference?, division: Division?): List<Team> =
-            teams.filter { conference == null || it.conference == conference }
-                    .filter { division == null || it.division == division }
-
-    override fun findTeam(name: String): Team = teams.first { it.name == name }
-}
