--- app/src/main/java/yt/javi/nfltweets/domain/model/Conference.kt
@@ -1,7 +1,0 @@
-package yt.javi.nfltweets.domain.model
-
-
-enum class Conference {
-    AFC,
-    NFC
-}
