--- app/src/main/java/yt/javi/nfltweets/domain/model/Division.kt
@@ -1,9 +1,0 @@
-package yt.javi.nfltweets.domain.model
-
-
-enum class Division {
-    East,
-    North,
-    South,
-    West
-}
