--- app/src/main/java/yt/javi/nfltweets/domain/model/team/TeamRepository.kt
@@ -1,11 +1,0 @@
-package yt.javi.nfltweets.domain.model.team
-
-import yt.javi.nfltweets.domain.model.Conference
-import yt.javi.nfltweets.domain.model.Division
-
-
-interface TeamRepository {
-    fun findTeams(conference: Conference?, division: Division?): List<Team>
-
-    fun findTeam(name: String): Team
-}
