+++ app/src/main/java/yt/javi/nftweets/domain/model/Conference.kt
@@ -1,0 +1,7 @@
+package yt.javi.nftweets.domain.model
+
+
+enum class Conference {
+    AFC,
+    NFC
+}
