+++ app/src/main/java/yt/javi/nftweets/domain/model/Division.kt
@@ -1,0 +1,9 @@
+package yt.javi.nftweets.domain.model
+
+
+enum class Division {
+    East,
+    North,
+    South,
+    West
+}
