+++ app/src/main/java/yt/javi/nftweets/domain/model/team/Team.kt
@@ -1,0 +1,14 @@
+package yt.javi.nftweets.domain.model.team
+
+import yt.javi.nftweets.domain.model.Conference
+import yt.javi.nftweets.domain.model.Division
+
+
+data class Team(
+        val name: String,
+        val city: String,
+        val conference: Conference,
+        val division: Division,
+        val logo: Int,
+        val twitterAccount: String
+)
