+++ app/src/main/java/yt/javi/nftweets/domain/model/team/TeamRepository.kt
@@ -1,0 +1,11 @@
+package yt.javi.nftweets.domain.model.team
+
+import yt.javi.nftweets.domain.model.Conference
+import yt.javi.nftweets.domain.model.Division
+
+
+interface TeamRepository {
+    fun findTeams(conference: Conference?, division: Division?): List<Team>
+
+    fun findTeam(name: String): Team
+}
