+++ app/src/main/java/yt/javi/nftweets/domain/service/team/GetTeamsByConferenceService.kt
@@ -1,0 +1,11 @@
+package yt.javi.nftweets.domain.service.team
+
+import yt.javi.nftweets.domain.model.Conference
+import yt.javi.nftweets.domain.model.team.Team
+import yt.javi.nftweets.domain.model.team.TeamRepository
+
+
+class GetTeamsByConferenceService(private val repository: TeamRepository) {
+    fun getTeamsByConference(conference: Conference): List<Team> =
+            repository.findTeams(conference, null)
+}
