+++ app/src/main/res/layout/activity_main.xml
@@ -1,13 +1,20 @@
 <?xml version="1.0" encoding="utf-8"?>
-<android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="yt.javi.nfltweets.MainActivity">
+    tools:context="yt.javi.nftweets.MainActivity"
+    android:orientation="vertical">
 
-    <android.support.v4.view.ViewPager
-        xmlns:android="http://schemas.android.com/apk/res/android"
+    <android.support.v7.widget.Toolbar
+        android:id="@+id/main_toolbar"
+        android:layout_width="match_parent"
+        android:layout_height="?attr/actionBarSize"
+        android:background="?attr/colorPrimary"
+        android:theme="@style/AppTheme.Toolbar" />
+
+    <android.support.v4.view.ViewPager xmlns:android="http://schemas.android.com/apk/res/android"
         android:id="@+id/pager"
         android:layout_width="match_parent"
         android:layout_height="wrap_content">
@@ -16,13 +23,14 @@
             android:id="@+id/tabs"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            app:tabMode="fixed"
             android:background="?attr/colorPrimary"
-            app:tabIndicatorHeight="12dp"
             app:tabGravity="fill"
-            app:tabTextColor="#FFF"
-            app:tabSelectedTextColor="@color/colorAccent" />
+            app:tabIndicatorColor="@color/primaryLightColor"
+            app:tabIndicatorHeight="6dp"
+            app:tabMode="fixed"
+            app:tabSelectedTextColor="@color/white"
+            app:tabTextColor="@color/white" />
 
-        </android.support.v4.view.ViewPager>
+    </android.support.v4.view.ViewPager>
 
-</android.support.constraint.ConstraintLayout>
+</LinearLayout>
