+++ app/src/main/res/values/colors.xml
@@ -1,8 +1,17 @@
-<?xml version="1.0" encoding="utf-8"?>
+<!--?xml version="1.0" encoding="UTF-8"?-->
 <resources>
-    <color name="colorPrimary">#3F51B5</color>
-    <color name="colorPrimaryDark">#303F9F</color>
-    <color name="colorAccent">#FF4081</color>
+    <color name="primaryColor">#013369</color>
+    <color name="colorPrimary">#013369</color>
+    <color name="primaryLightColor">#405c97</color>
+    <color name="primaryDarkColor">#000c3e</color>
+    <color name="colorPrimaryDark">#000c3e</color>
+    <color name="secondaryColor">#ffffff</color>
+    <color name="colorAccent">#ffffff</color>
+    <color name="secondaryLightColor">#ffffff</color>
+    <color name="secondaryDarkColor">#cccccc</color>
+    <color name="primaryTextColor">#ffffff</color>
+    <color name="secondaryTextColor">#000000</color>
     <color name="black_alpha_30">#4b000000</color>
     <color name="transparent">#00000000</color>
+    <color name="white">#ffffff</color>
 </resources>
