--- app/src/test/java/yt/javi/nfltweets/ExampleUnitTest.kt
@@ -1,17 +1,0 @@
-package yt.javi.nfltweets
-
-import org.junit.Assert.assertEquals
-import org.junit.Test
-
-/**
- * Example local unit test, which will execute on the development machine (host).
- *
- * @see [Testing documentation](http://d.android.com/tools/testing)
- */
-class ExampleUnitTest {
-    @Test
-    @Throws(Exception::class)
-    fun addition_isCorrect() {
-        assertEquals(4, (2 + 2).toLong())
-    }
-}
